# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: translator-server API
  version: 1.0.0
  summary: Canonical REST API for the Yara translator-server.
  description: |
    The canonical API surface is mounted under `/api/v1`. Legacy aliases under
    `/api/db/*`, `/api/user/*`, `/api/epubs/*`, `/api/backup/*`,
    `/api/browser-workers`, `/api/proxy/*`, `/api/defaults`,
    `/api/translation-jobs/*` are **deprecated** and will be removed on
    **Wed, 01 Jan 2026** (the `Sunset` header is set on every legacy response).
    New clients must use `/api/v1`.

    When the running server and this document diverge, the running server is the
    source of truth.
  contact:
    name: Yara
    url: https://github.com/mfloresz/yara
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /api/v1
    description: Canonical v1 surface
  - url: /
    description: Server root (legacy aliases live directly under /api/...)
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /prompts/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Prompts]
      summary: Create or update a prompt.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PromptUpsert' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Prompt' }

  # ---------------------------------------------------------------------------
  # Providers
  # ---------------------------------------------------------------------------
...
  /providers/{providerKey}:
    parameters:
      - name: providerKey
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Providers]
      summary: Update model, baseUrl, timeoutMs, concurrency.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderUpdate' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderSettings' }
  /providers/{providerKey}/key:
    parameters:
      - name: providerKey
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Providers]
      summary: Replace the API key (write-only, encrypted at rest).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [apiKey]
              properties:
                apiKey: { type: string, format: password }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProviderSettings' }
    delete:
      tags: [Providers]
      summary: Delete the stored API key.
      responses:
        '204': { $ref: '#/components/responses/NoContent' }

  # ---------------------------------------------------------------------------
  # Settings
  # ---------------------------------------------------------------------------
...
  /novels/{novelId}/reading-progress:
    parameters:
      - $ref: '#/components/parameters/NovelIdParam'
    get:
      tags: [Reading progress]
      summary: Get the user's last-read chapter and scroll position.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ReadingProgress' }
    put:
      tags: [Reading progress]
      summary: Save reading progress.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chapterId, scrollPercent]
              properties:
                chapterId: { type: string }
                scrollPercent: { type: number, format: float, minimum: 0, maximum: 1 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ReadingProgress' }

  # ---------------------------------------------------------------------------
  # Imports / downloads
  # ---------------------------------------------------------------------------
...
  /proxy/fetch:
    post:
      tags: [Browser workers]
      summary: Fetch a URL through the connected browser worker.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
                timeout: { type: integer, default: 120, maximum: 300 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ProxyFetchResult' }
        '400':
          description: 'No browser worker connected, or timeout.'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ---------------------------------------------------------------------------
  # Worker auth
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth.token
  parameters:
...
    NovelIdParam:
      name: novelId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    NoContent:
      description: No Content.
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem details. The envelope uses `error` instead of the standard top-level fields to keep the v1 envelope consistent.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, unauthorized, forbidden, not_found, conflict, queue_full, internal_error]
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
                  code: { type: string }
    SingleEnvelope:
      type: object
      description: Canonical v1 single-resource envelope.
      required: [data]
      properties:
        data: {}
...
    Prompt:
      type: object
      required: [id, key]
      properties:
        id: { type: string, description: Same as `key`. }
        key: { type: string }
        label: { type: string }
        description: { type: string }
        prompt:
          type: object
          properties:
            systemPrompt: { type: string }
            userPrompt: { type: string }
        active: { type: boolean }
        updatedAt: { type: string }
    PromptUpsert:
      type: object
      properties:
        label: { type: string }
        description: { type: string }
        prompt:
          type: object
          properties:
            systemPrompt: { type: string }
            userPrompt: { type: string }
        active: { type: boolean }
    ProviderSettings:
      type: object
      required: [key]
      properties:
        key: { type: string }
        model: { type: string }
        baseUrl: { type: string }
        timeoutMs: { type: integer }
        concurrency: { type: integer, minimum: 1, maximum: 10, default: 1 }
        apiKeyConfigured: { type: boolean, description: True if a key is on file. The key itself is never returned. }
    ProviderUpdate:
      type: object
      properties:
        model: { type: string }
        baseUrl: { type: string }
        timeoutMs: { type: integer }
        concurrency: { type: integer, minimum: 1, maximum: 10 }
...
    ReadingProgress:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        novelId: { type: string }
        chapterId: { type: string }
        scrollPercent: { type: number, format: float, minimum: 0, maximum: 1 }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    ProxyFetchResult:
      type: object
      properties:
        url: { type: string, format: uri }
        title: { type: string }
        html: { type: string }
        text: { type: string }
        status: { type: string, example: ok }
...